Lavorazione chord ml
