--- config/shapes/programs/27ac52ad6baf4e86b35457880cf5049e/state.json
@@ -1,47 +1,0 @@
-{
-  "blocks": {
-    "languageVersion": 0,
-    "blocks": [
-      {
-        "type": "tactigon_shape_setup",
-        "id": "{HLjiI#GI%%_dH1=`/(j",
-        "x": 30,
-        "y": 30,
-        "deletable": false
-      },
-      {
-        "type": "tactigon_shape_function",
-        "id": "Vva^SLWL_2oC]KIH-TWo",
-        "x": 30,
-        "y": 150,
-        "deletable": false,
-        "inputs": {
-          "BODY": {
-            "block": {
-              "type": "tactigon_shape_debug",
-              "id": "aX31Z~8k.~5Tg~hh87Az",
-              "inputs": {
-                "TEXT": {
-                  "shadow": {
-                    "type": "text",
-                    "id": "o4%A25]q8+6P2);ugM/J",
-                    "fields": {
-                      "TEXT": "Tactigon"
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      },
-      {
-        "type": "tactigon_shape_close",
-        "id": "/]|yP}j*a10:iVzU#.)k",
-        "x": 30,
-        "y": 300,
-        "deletable": false
-      }
-    ]
-  }
-}
